Address comment and improve testcase
