+++ core/src/main/java/org/apache/iceberg/avro/TypeToSchema.java
@@ -19,6 +19,7 @@
 
 package org.apache.iceberg.avro;
 
+import com.google.common.base.Preconditions;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import java.util.List;
@@ -87,13 +88,13 @@ class TypeToSchema extends TypeUtil.SchemaVisitor<Schema> {
     for (int i = 0; i < structFields.size(); i += 1) {
       Types.NestedField structField = structFields.get(i);
       String origFieldName = structField.name();
-      boolean isValidFieldName = validAvroName(origFieldName);
-      String fieldName =  isValidFieldName ? origFieldName : sanitize(origFieldName);
+      boolean isValidFieldName = AvroSchemaUtil.validAvroName(origFieldName);
+      String fieldName =  isValidFieldName ? origFieldName : AvroSchemaUtil.sanitize(origFieldName);
       Schema.Field field = new Schema.Field(
           fieldName, fieldSchemas.get(i), null,
           structField.isOptional() ? JsonProperties.NULL_VALUE : null);
       if (!isValidFieldName) {
-        field.addProp(AvroSchemaUtil.ORIGINAL_FIELD_NAME_PROP, origFieldName);
+        field.addProp(AvroSchemaUtil.ICEBERG_FIELD_NAME_PROP, origFieldName);
       }
       field.addProp(AvroSchemaUtil.FIELD_ID_PROP, structField.fieldId());
       fields.add(field);
@@ -219,47 +220,5 @@ class TypeToSchema extends TypeUtil.SchemaVisitor<Schema> {
     results.put(primitive, primitiveSchema);
 
     return primitiveSchema;
-  }
-
-  private static String sanitize(String name) {
-    int length = name.length();
-    StringBuilder sb = new StringBuilder(name.length());
-    char first = name.charAt(0);
-    if (!(Character.isLetter(first) || first == '_')) {
-      sb.append(sanitize(first));
-    } else {
-      sb.append(first);
-    }
-    for (int i = 1; i < length; i++) {
-      char character = name.charAt(i);
-      if (!(Character.isLetterOrDigit(character) || character == '_')) {
-        sb.append(sanitize(character));
-      } else {
-        sb.append(character);
-      }
-    }
-    return sb.toString();
-  }
-
-  private static String sanitize(char character) {
-    return "_x" + Integer.toHexString(character).toUpperCase();
-  }
-
-  private static boolean validAvroName(String name) {
-    int length = name.length();
-    if (length == 0) {
-      throw new IllegalArgumentException("Empty name " + name);
-    }
-    char first = name.charAt(0);
-    if (!(Character.isLetter(first) || first == '_')) {
-      return false;
-    }
-    for (int i = 1; i < length; i++) {
-      char character = name.charAt(i);
-      if (!(Character.isLetterOrDigit(character) || character == '_')) {
-        return false;
-      }
-    }
-    return true;
   }
 }
